Move restlib to pkg folder
